pyproject.toml
[project]
name = "car-sales-data-visualisation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "kaggle>=2.0.0",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
]
